+++ src/db/app_catalog_cache.rs
@@ -8,7 +8,7 @@ use tracing::{info, warn};
 /// performance bottleneck during high-volume query log insertion.
 pub struct AppCatalogCache {
     /// Maps a raw domain string (e.g., "youtube.com") to its App ID.
-    domains_to_app_id: RwLock<HashMap<String, i64>>,
+    domains_to_app_id: RwLock<HashMap<String, i32>>,
 }
 
 impl Default for AppCatalogCache {
@@ -27,7 +27,7 @@ impl AppCatalogCache {
 
     /// Load the domain-to-app mappings from the database.
     pub async fn load_from_db(&self, db: &DbPool) {
-        let rows: Result<Vec<(i64, String)>, sqlx::Error> =
+        let rows: Result<Vec<(i32, String)>, sqlx::Error> =
             sqlx::query_as("SELECT app_id, domain FROM app_domains")
                 .fetch_all(db)
                 .await;
@@ -51,7 +51,7 @@ impl AppCatalogCache {
     /// Match a domain against the loaded catalog.
     /// Checks exact match first, then checks successive parent domains
     /// (e.g., sub.test.example.com -> test.example.com -> example.com)
-    pub fn match_domain(&self, domain: &str) -> Option<i64> {
+    pub fn match_domain(&self, domain: &str) -> Option<i32> {
         let cache = self.domains_to_app_id.read().unwrap();
         if cache.is_empty() {
             return None;
